requirements.txt
python-dotenv
ringcentral
pydash
boto3
flask
ringcentral_bot_framework